Fixed issue with ObjMetadata in object search permission service test
- ObjectMetadata is now updated if a row exists
