+++ packages/grid/backend/grid/db/session.py
@@ -22,7 +22,6 @@ def get_db_engine(db_uri: str = str(settings.SQLALCHEMY_DATABASE_URI)) -> Engine
     :rtype: Engine
     """
     if db_uri.startswith("sqlite://"):
-
         db_engine = create_engine(db_uri, echo=False)
         # TODO change to use alembic properly with the sqlite memory store:
         # https://stackoverflow.com/questions/31406359/use-alembic-to-upgrade-in-memory-sqlite3-database
