+++ packages/syft/src/syft/core/adp/vm_private_scalar_manager.py
@@ -47,12 +47,12 @@ class PrimeFactory:
 
     def __init__(self) -> None:
         self.exp = 2
-        self.prime_numbers: list = get_cached_primes(total=10 ** self.exp)
+        self.prime_numbers: list = get_cached_primes(total=10**self.exp)
 
     def get(self, index: int) -> int:
         while index > len(self.prime_numbers) - 1:
             self.exp += 1
-            self.prime_numbers = get_cached_primes(total=10 ** self.exp)
+            self.prime_numbers = get_cached_primes(total=10**self.exp)
         return self.prime_numbers[index]
 
     def __eq__(self, other: Any) -> bool:
