+++ packages/syft/tests/syft/core/node/domain/service/request_message_test.py
@@ -13,7 +13,7 @@ from syft import deserialize
 from syft import serialize
 from syft.core.common import UID
 from syft.core.io.address import Address
-from syft.core.node.common.node_manager.bin_obj_manager import DictStore
+from syft.core.node.common.node_manager.dict_store import DictStore
 from syft.core.node.common.node_service.request_receiver.request_receiver_messages import (
     RequestMessage,
 )
