+++ packages/syft/tests/syft/core/tensor/mnist_test.py
@@ -8,7 +8,7 @@ import syft as sy
 from syft import deserialize
 from syft import serialize
 from syft.core.adp.entity import Entity
-from syft.core.node.common.node_manager.bin_obj_manager import DictStore
+from syft.core.node.common.node_manager.dict_store import DictStore
 from syft.core.tensor.tensor import Tensor
